closure: fix final persistence audit items (slot-15 dispatch, host guards, companion path, stream type, test isolation)
INPUT_LoadSaveFile slot-15 child dispatch now models both per-class
semantics with evidence: ResourceGameObject slot [15] = CreateMember
(0x458430) creates/attaches a member Building, while Building slot [15]
= Update(void*) (0x4327B0, AI dispatch) — the binary's result null-check
at 0x41D7F8 would treat a non-zero field_dc as a bogus child (binary
defect; null path invariant on load).  Typed member dispatch, no raw
vtable access: host RTTI dynamic_cast behind #ifndef _WIN32, _WIN32 uses
the canonical type tag (type=3 from ResourceGameObject ctor).

Scenario-loop GetItem and SaveCurrentWorld resource/name-owner null
hardenings moved inside #ifndef _WIN32; _WIN32 keeps the binary's
unconditional derefs, each documented with its address.  INPUT_LoadWorld
_WIN32 companion now replaces the backdrop save-name's last four
characters with ".sav" (0x41D3C9/0x41D3CC) instead of appending; host
keeps the documented curr.sav snapshot/restore.  ResDataSave
stream_byte_count widened to uint32_t (fixes 64-bit-host
-Wsign-compare).

GUI integration no longer points LEGO_LOCO_DATA at source art-res:
session-scoped game_data_dir fixture copies the whole data root
(Exe/loco.exe for the PE string table, ghidra_projects excluded) into
build/test-artifacts; gui_sandbox GameSession exports data_dir.
Removed the generated untracked art-res/curr artifact.  Verified: make
(135/135), make check, make test (26 PASS), make test-integration
(11/11), _WIN32 syntax checks.

+++ src/decompiled_cpp/resources/ResDataSave.cpp
@@ -310,9 +310,15 @@ WIN32_StreamDestroyImmediate(stream);
  * conflated.  Typed stream reconstruction is deferred (TODO: WIN32_Stream
  * class during the stream-integration milestone); these helpers
  * reproduce the exact raw reads. */
-static int32_t stream_byte_count(void* stream)
+static uint32_t stream_byte_count(void* stream)
 {
-    return *reinterpret_cast<int32_t*>(static_cast<uint8_t*>(stream) + 8);
+    /* The last-read/written byte count is a signed dword in the binary
+     * (0x447DD1 cmpl $0x80,0x8(%ecx) — signed 32-bit compare); the
+     * count is never negative in practice.  Return it as uint32_t so
+     * the short-read comparisons against sizeof/preview sizes below
+     * are unsigned-to-unsigned (no -Wsign-compare on 64-bit hosts). */
+    return static_cast<uint32_t>(
+        *reinterpret_cast<int32_t*>(static_cast<uint8_t*>(stream) + 8));
 }
 
 static int32_t stream_state_word(void* stream)
